pyproject.toml
[project]
name = "expense-tracker-bot"
version = "0.1.0"
description = "Expense tracker bot with Telegram integration"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-telegram-bot>=20.7",
    "openai>=1.3.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.108.0",
    "apscheduler>=3.10.0",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "pymongo>=4.6.0", # Optional, for chat history
    "gspread>=6.1.4",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
]
